Improve handling of too many spans error.
Previously, when the max spans were exceeded, we'd mark the trace as
broken and not attempt to report it. With
https://github.com/tildeio/skylight-rust/pull/56 the Rust agent will
now report a truncated trace to the collector instead.

This PR allows opting-in to that new behavior by setting
`report_max_spans_exceeded` in the config. In general, we recommend
doing this but recognize that some users may not appreciate seeing
additional requests logged without the full event sequence.